pyproject.toml
[project]
name = "parsl-saga-integration"
version = "0.1.0"
description = "SAGA + Parsl Integration for Advanced Task Scheduling"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Jared Coleman", email = "jared@example.com"}
]
dependencies = [
    "parsl",
    "anrg-saga",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "mypy>=0.991",
]

[tool.uv.sources]
# Local editable installations of submodules
parsl = { path = "./parsl", editable = true }
anrg-saga = { path = "./saga", editable = true }
